[frontend] move test next to source (best practice starting now).
