+++ opencti-platform/opencti-front/setup-relay-for-vitest.ts
